+++ examples/snap-chapi-example/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @learncard/snap-chapi-example
 
+## 1.0.187
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/core@9.3.8
+    -   @learncard/helpers@1.0.20
+    -   @learncard/react@2.7.43
+    -   @learncard/meta-mask-snap@1.0.59
+
+## 1.0.186
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.7.42
+
+## 1.0.185
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/react@2.7.41
+
 ## 1.0.184
 
 ### Patch Changes