+++ packages/learn-card-init/CHANGELOG.md
@@ -1,5 +1,51 @@
 # learn-card-core
 
+## 1.3.14
+
+### Patch Changes
+
+-   Updated dependencies [[`72e9661ffe0c9f9e3c312ecba2b6441d61941a4a`](https://github.com/learningeconomy/LearnCard/commit/72e9661ffe0c9f9e3c312ecba2b6441d61941a4a)]:
+    -   @learncard/network-plugin@2.1.8
+    -   @learncard/types@5.5.8
+    -   @learncard/core@9.3.8
+    -   @learncard/helpers@1.0.20
+    -   @learncard/ceramic-plugin@1.0.19
+    -   @learncard/chapi-plugin@1.0.35
+    -   @learncard/did-web-plugin@1.0.45
+    -   @learncard/didkey-plugin@1.0.19
+    -   @learncard/didkit-plugin@1.4.17
+    -   @learncard/idx-plugin@1.0.19
+    -   @learncard/learn-card-plugin@1.1.21
+    -   @learncard/learn-cloud-plugin@2.0.23
+    -   @learncard/vc-plugin@1.1.22
+    -   @learncard/vc-api-plugin@1.0.19
+    -   @learncard/vc-templates-plugin@1.0.29
+    -   @learncard/vpqr-plugin@1.0.19
+    -   @learncard/expiration-plugin@1.1.22
+    -   @learncard/crypto-plugin@1.0.19
+    -   @learncard/dynamic-loader-plugin@1.0.15
+    -   @learncard/ethereum-plugin@1.0.20
+
+## 1.3.13
+
+### Patch Changes
+
+-   Updated dependencies [[`df13a7de5f31fabd842884a3e7c0706746c60da1`](https://github.com/learningeconomy/LearnCard/commit/df13a7de5f31fabd842884a3e7c0706746c60da1)]:
+    -   @learncard/vc-templates-plugin@1.0.28
+    -   @learncard/learn-cloud-plugin@2.0.22
+    -   @learncard/did-web-plugin@1.0.44
+    -   @learncard/network-plugin@2.1.7
+
+## 1.3.12
+
+### Patch Changes
+
+-   Updated dependencies [[`9afd9e515d1cefeac03f76fb13556e6288abbc35`](https://github.com/learningeconomy/LearnCard/commit/9afd9e515d1cefeac03f76fb13556e6288abbc35)]:
+    -   @learncard/vc-templates-plugin@1.0.27
+    -   @learncard/did-web-plugin@1.0.43
+    -   @learncard/learn-cloud-plugin@2.0.21
+    -   @learncard/network-plugin@2.1.6
+
 ## 1.3.11
 
 ### Patch Changes