+++ packages/react-learn-card/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @welibraryos/react-learn-card
 
+## 2.7.43
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.3.14
+
+## 2.7.42
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.3.13
+
+## 2.7.41
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.3.12
+
 ## 2.7.40
 
 ### Patch Changes